Fix size of invisible container for floating buttons
The floating buttons containers were too large and were hovering on top
of other UI elements. When you tried to click the UI elements, they
wouldn't work. It was caused by the invisible container hovering on top
of those elements on a higher z-index, thus catching the click event and
making it seem like the UI element didn't work.